Add files via upload
Question of Interest: Do the possible predictor variables pertaining to size of the house, specifically the number of
beds and baths, square feet, lot size, garage size, and the presence of a pool, or lack thereof,
affect the sales price of the homes?

Data Source: The real estate data was assembled by a city tax assessor, and comprises the data of 521
transactions in the residential home sale industry in a midwestern city during the year of 2002. It
keeps a record of 12 variables in terms of price, size, architectural soundness, quality, and
amenities. We will take into consideration some of these variables in terms of how they affect
the sales price of residences.